Extract SiteMediaCollectionViewController
