+++ WordPress/Classes/ViewRelated/Media/SiteMedia/Helpers/MediaCollectionCell.swift
