+++ WordPress/Classes/ViewRelated/Media/SiteMedia/Helpers/MediaCollectionCellBadgeView.swift
