+++ WordPress/Classes/ViewRelated/Media/SiteMedia/Helpers/MediaCollectionCellViewModel.swift
